requirements.txt
pip == 23.3.1
Django == 4.2.6
djangorestframework == 3.14.0
psycopg2 == 2.9.9
djangorestframework-simplejwt == 5.3.0
django-crispy-forms == 2.1
crispy-bootstrap5 == 0.7



